requirements.txt
# Web App
Flask==3.0.3
requests==2.32.2

# Content
google-generativeai==0.8.2
